refactor(web): extract idle watchdog from stream-transport into stream-watchdog
- Extract watchdog timer, liveness counters, and Sentry telemetry into
  createStreamWatchdog() factory in stream-watchdog.ts (208 lines)
- stream-transport.ts drops from 488 to 289 lines (under ~300 guideline)
- Shared types (ChatEventStream, ChatStreamReconnectCause, ChatEventStreamOptions)
  moved to stream-transport-types.ts to break circular dependency
- Move 4 unique polling/SSE race condition tests from stream-transport.test.ts
  to turn-store.test.ts (they test turnReducer, not stream transport)
- Drop 8 duplicate tests already covered by turn-store.test.ts

Closes LUM-2056
