+++ apps/web/src/lib/streaming/stream-transport.test.ts
@@ -46,259 +46,8 @@ mock.module("@sentry/browser", () => ({
 import {
   getDiagnosticsEvents,
 } from "@/lib/diagnostics";
-import {
-  type TurnState,
-  INITIAL_TURN_STATE,
-  turnReducer,
-  isSending,
-} from "@/domains/chat/turn-store";
-import { parseAssistantEvent } from "@/lib/streaming/event-parser";
 import { subscribeChatEvents, type ChatStreamReconnectCause } from "@/lib/streaming/stream-transport";
 import { useAssistantIdentityStore } from "@/stores/assistant-identity-store";
-
-describe("polling reconciliation with state machine", () => {
-  /**
-   * Simulates the page-level flow where polling completion dispatches a
-   * POLL_RECONCILED event through the turn reducer.  These tests verify
-   * the contract between the polling path and the state machine.
-   */
-
-  test("poll completion transitions active turn to idle", () => {
-    // Simulate: user sends (thinking) -> poll finds reply -> dispatch POLL_RECONCILED
-    const afterSend = turnReducer(INITIAL_TURN_STATE, {
-      type: "USER_SEND_REQUESTED",
-      turnId: "t-poll-1",
-    });
-    expect(isSending(afterSend)).toBe(true);
-
-    const afterPoll = turnReducer(afterSend, {
-      type: "POLL_RECONCILED",
-      turnId: "t-poll-1",
-    });
-    expect(afterPoll.phase).toBe("idle");
-    expect(isSending(afterPoll)).toBe(false);
-    expect(afterPoll.lastTerminalReason).toBe("complete");
-  });
-
-  test("SSE completes before poll — poll for same turnId is no-op", () => {
-    // SSE path: send -> delta -> message_complete
-    let state = turnReducer(INITIAL_TURN_STATE, {
-      type: "USER_SEND_REQUESTED",
-      turnId: "t-race-1",
-    });
-    state = turnReducer(state, { type: "ASSISTANT_TEXT_DELTA" });
-    state = turnReducer(state, { type: "MESSAGE_COMPLETE" });
-    expect(state.phase).toBe("idle");
-    expect(state.activeTurnId).toBeNull();
-
-    // Poll arrives late for the same turn — should be a no-op
-    const afterPoll = turnReducer(state, {
-      type: "POLL_RECONCILED",
-      turnId: "t-race-1",
-    });
-    expect(afterPoll).toEqual(state);
-  });
-
-  test("poll completes before SSE — SSE message_complete is idempotent", () => {
-    // Poll path completes first
-    let state = turnReducer(INITIAL_TURN_STATE, {
-      type: "USER_SEND_REQUESTED",
-      turnId: "t-race-2",
-    });
-    state = turnReducer(state, {
-      type: "POLL_RECONCILED",
-      turnId: "t-race-2",
-    });
-    expect(state.phase).toBe("idle");
-
-    // SSE message_complete arrives late — already idle, idempotent
-    const afterSSE = turnReducer(state, { type: "MESSAGE_COMPLETE" });
-    expect(afterSSE.phase).toBe("idle");
-    expect(afterSSE.lastTerminalReason).toBe("complete");
-  });
-
-  test("stale poll does not interfere with new turn", () => {
-    // Turn 1: send -> complete
-    let state = turnReducer(INITIAL_TURN_STATE, {
-      type: "USER_SEND_REQUESTED",
-      turnId: "t-old",
-    });
-    state = turnReducer(state, { type: "MESSAGE_COMPLETE" });
-
-    // Turn 2: send (now active)
-    state = turnReducer(state, {
-      type: "USER_SEND_REQUESTED",
-      turnId: "t-new",
-    });
-    expect(state.activeTurnId).toBe("t-new");
-    expect(state.phase).toBe("thinking");
-
-    // Stale poll from turn 1 arrives — should NOT affect turn 2
-    const afterStalePoll = turnReducer(state, {
-      type: "POLL_RECONCILED",
-      turnId: "t-old",
-    });
-    expect(afterStalePoll.phase).toBe("thinking");
-    expect(afterStalePoll.activeTurnId).toBe("t-new");
-  });
-
-  test("poll without turnId still works for backward compatibility", () => {
-    const afterSend = turnReducer(INITIAL_TURN_STATE, {
-      type: "USER_SEND_REQUESTED",
-      turnId: "t-compat",
-    });
-    // Legacy-style poll without turnId
-    const afterPoll = turnReducer(afterSend, {
-      type: "POLL_RECONCILED",
-    });
-    expect(afterPoll.phase).toBe("idle");
-    expect(afterPoll.lastTerminalReason).toBe("complete");
-  });
-
-  test("SSE events mapped from wire format produce correct domain events", () => {
-    // Verify that parseAssistantEvent produces events that map correctly
-    // to domain events consumed by the reducer
-    const delta = parseAssistantEvent({
-      type: "assistant_text_delta",
-      text: "hello",
-    });
-    expect(delta.type).toBe("assistant_text_delta");
-
-    const complete = parseAssistantEvent({
-      type: "message_complete",
-      messageId: "msg-1",
-    });
-    expect(complete.type).toBe("message_complete");
-
-    const handoff = parseAssistantEvent({
-      type: "generation_handoff",
-      queuedCount: 0,
-    });
-    expect(handoff.type).toBe("generation_handoff");
-
-    const error = parseAssistantEvent({ type: "error", message: "fail" });
-    expect(error.type).toBe("error");
-
-    // The page maps these wire types to domain event types:
-    // assistant_text_delta -> ASSISTANT_TEXT_DELTA
-    // message_complete     -> MESSAGE_COMPLETE
-    // generation_handoff   -> GENERATION_HANDOFF
-    // error                -> STREAM_ERROR
-
-    // Verify domain events flow correctly through reducer
-    let state: TurnState = turnReducer(INITIAL_TURN_STATE, {
-      type: "USER_SEND_REQUESTED",
-      turnId: "t-wire",
-    });
-    state = turnReducer(state, { type: "ASSISTANT_TEXT_DELTA" });
-    expect(state.phase).toBe("streaming");
-    state = turnReducer(state, { type: "MESSAGE_COMPLETE" });
-    expect(state.phase).toBe("idle");
-  });
-
-  // ---- activeTurnId guard on idle re-activation ----
-
-  test("ASSISTANT_TEXT_DELTA does NOT re-activate idle when activeTurnId is null", () => {
-    let state = turnReducer(INITIAL_TURN_STATE, {
-      type: "USER_SEND_REQUESTED",
-      turnId: "t-guard-1",
-    });
-    state = turnReducer(state, { type: "MESSAGE_COMPLETE" });
-    expect(state.phase).toBe("idle");
-    expect(state.activeTurnId).toBeNull();
-
-    const afterDelta = turnReducer(state, { type: "ASSISTANT_TEXT_DELTA" });
-    expect(afterDelta.phase).toBe("idle");
-  });
-
-  test("ASSISTANT_TEXT_DELTA re-activates idle when activeTurnId is set", () => {
-    const forcedIdle: TurnState = {
-      phase: "idle",
-      activeTurnId: "t-guard-2",
-      activeToolCallCount: 0,
-      pendingQueuedCount: 0,
-      lastTerminalReason: null,
-      statusText: null,
-      liveWebActivity: {},
-      autoRoutedProfileLabel: null,
-    };
-    const afterDelta = turnReducer(forcedIdle, { type: "ASSISTANT_TEXT_DELTA" });
-    expect(afterDelta.phase).toBe("streaming");
-    expect(afterDelta.activeTurnId).toBe("t-guard-2");
-  });
-
-  test("TOOL_USE_START does NOT re-activate idle when activeTurnId is null", () => {
-    let state = turnReducer(INITIAL_TURN_STATE, {
-      type: "USER_SEND_REQUESTED",
-      turnId: "t-guard-3",
-    });
-    state = turnReducer(state, { type: "MESSAGE_COMPLETE" });
-    expect(state.phase).toBe("idle");
-    expect(state.activeTurnId).toBeNull();
-
-    const afterTool = turnReducer(state, { type: "TOOL_USE_START" });
-    expect(afterTool.phase).toBe("idle");
-    expect(afterTool.activeToolCallCount).toBe(0);
-  });
-
-  test("TOOL_USE_START re-activates idle when activeTurnId is set", () => {
-    const forcedIdle: TurnState = {
-      phase: "idle",
-      activeTurnId: "t-guard-4",
-      activeToolCallCount: 0,
-      pendingQueuedCount: 0,
-      lastTerminalReason: null,
-      statusText: null,
-      liveWebActivity: {},
-      autoRoutedProfileLabel: null,
-    };
-    const afterTool = turnReducer(forcedIdle, { type: "TOOL_USE_START" });
-    expect(afterTool.phase).toBe("thinking");
-    expect(afterTool.activeToolCallCount).toBe(1);
-  });
-
-  test("POLL_RECONCILED → stale delta → new send works end-to-end", () => {
-    // Simulates the full background/foreground race:
-    // 1. Turn starts → streaming
-    // 2. POLL_RECONCILED idles the turn (activeTurnId → null)
-    // 3. Stale ASSISTANT_TEXT_DELTA arrives → should be ignored
-    // 4. User sends new message → new turn starts cleanly
-    let state = turnReducer(INITIAL_TURN_STATE, {
-      type: "USER_SEND_REQUESTED",
-      turnId: "t-race-bg",
-    });
-    state = turnReducer(state, { type: "ASSISTANT_TEXT_DELTA" });
-    expect(state.phase).toBe("streaming");
-
-    state = turnReducer(state, {
-      type: "POLL_RECONCILED",
-      turnId: "t-race-bg",
-    });
-    expect(state.phase).toBe("idle");
-    expect(state.activeTurnId).toBeNull();
-
-    state = turnReducer(state, { type: "ASSISTANT_TEXT_DELTA" });
-    expect(state.phase).toBe("idle");
-
-    state = turnReducer(state, {
-      type: "USER_SEND_REQUESTED",
-      turnId: "t-race-new",
-    });
-    expect(state.phase).toBe("thinking");
-    expect(state.activeTurnId).toBe("t-race-new");
-  });
-
-  test("ASSISTANT_TEXT_DELTA still transitions thinking → streaming (no guard)", () => {
-    let state = turnReducer(INITIAL_TURN_STATE, {
-      type: "USER_SEND_REQUESTED",
-      turnId: "t-thinking",
-    });
-    expect(state.phase).toBe("thinking");
-
-    state = turnReducer(state, { type: "ASSISTANT_TEXT_DELTA" });
-    expect(state.phase).toBe("streaming");
-  });
-});
 
 describe("subscribeChatEvents idle watchdog", () => {
   let originalFetch: typeof fetch;
