+++ apps/web/src/lib/streaming/stream-transport.ts
@@ -2,15 +2,13 @@
  * SSE stream transport for real-time assistant events.
  *
  * Opens an EventSource-style connection to the daemon's events endpoint,
- * automatically reconnects with exponential backoff, and includes an idle
- * watchdog to detect silently stalled connections (notably on iOS WKWebView).
+ * automatically reconnects with exponential backoff, and delegates idle
+ * detection to {@link createStreamWatchdog} to catch silently stalled
+ * connections (notably on iOS WKWebView).
  */
-
-import * as Sentry from "@sentry/browser";
 
 import { client } from "@/generated/api/client.gen";
 import { SDK_BASE_OPTIONS } from "@/utils/api-errors";
-import { recordDiagnostic, resolvePlatformTag } from "@/lib/diagnostics";
 import { parseAssistantEvent } from "@/lib/streaming/event-parser";
 import type { AssistantEvent } from "@/types/event-types";
 import { pickConversationIdWireField } from "@/lib/backwards-compat/conversation-id-wire-field";
@@ -21,6 +19,18 @@ import {
   registerSseClient,
   unregisterSseClient,
 } from "@/lib/streaming/stream-debug";
+import { createStreamWatchdog } from "@/lib/streaming/stream-watchdog";
+
+export type {
+  ChatEventStream,
+  ChatEventStreamOptions,
+  ChatStreamReconnectCause,
+} from "@/lib/streaming/stream-transport-types";
+import type {
+  ChatEventStream,
+  ChatEventStreamOptions,
+  ChatStreamReconnectCause,
+} from "@/lib/streaming/stream-transport-types";
 
 // ---------------------------------------------------------------------------
 // SSE stream transport
@@ -36,70 +46,6 @@ const STREAM_MAX_RECONNECT_DELAY_MS = 30_000;
 // must comfortably exceed that interval to avoid false positives on a
 // healthy connection that is idle between user turns.
 const STREAM_IDLE_TIMEOUT_MS = 45_000;
-
-export interface ChatEventStream {
-  /** Cancel the stream. Safe to call multiple times. */
-  cancel: () => void;
-}
-
-/**
- * Why the previous connection attempt was torn down. `"error"` covers
- * the SDK surfacing a fetch failure or the iterator ending; `"watchdog"`
- * means the client-side idle timer fired because no SSE traffic
- * (events or heartbeat comments) arrived within the configured window.
- * Threaded through to {@link ChatEventStreamOptions.onReconnect} so
- * callers can distinguish silent-stall recoveries from ordinary
- * transport errors when recording telemetry.
- */
-export type ChatStreamReconnectCause = "error" | "watchdog";
-
-export interface ChatEventStreamOptions {
-  /**
-   * Called after the SSE transport successfully reconnects. The events
-   * endpoint is live-only, so callers should use this hook to reconcile
-   * authoritative conversation history for messages emitted while
-   * offline. The `cause` argument indicates whether the previous attempt
-   * ended via a transport error or because the idle watchdog fired.
-   */
-  onReconnect?: (cause: ChatStreamReconnectCause) => void | Promise<void>;
-  /**
-   * Maximum interval, in milliseconds, with no SSE traffic from the
-   * server (events OR heartbeat comments) before the client treats the
-   * stream as silently stalled and force-reconnects.
-   *
-   * The fetch promise on iOS WKWebView (Capacitor) and some intermediate
-   * proxies can hold a streaming connection open at the network layer
-   * while no bytes flow through, with no error surfaced to JavaScript.
-   * Without a client-side liveness check, the stream sits forever
-   * waiting on the next byte. Defaults to {@link STREAM_IDLE_TIMEOUT_MS}.
-   * Mainly exposed for tests.
-   */
-  idleTimeoutMs?: number;
-  /**
-   * Base delay, in milliseconds, used by the exponential-backoff
-   * scheduler before the next reconnect attempt after a stream drop or
-   * a watchdog-driven stall. Mainly exposed for tests.
-   */
-  reconnectBaseDelayMs?: number;
-  /**
-   * Snapshot whether the caller-owned turn state machine is currently in
-   * a sending phase. When provided, the result is forwarded to Sentry on
-   * watchdog fires as the `wasTurnSending` tag and extra so the
-   * `sse_watchdog_fired` event count can be split into
-   * user-harming (`true`: a stall while the user is waiting for an
-   * in-flight assistant turn) vs benign (`false`: a stall on an idle
-   * stream after a turn completed). Without this split, the 100%
-   * `messagesAddedBucket=0` fleet reading collapses both populations and
-   * is uninterpretable for the Layer 2 / Layer 3 decision. Optional;
-   * defaults to omitting the tag entirely (Sentry treats absent tags as
-   * `"<absent>"` in Discover grouping).
-   *
-   * Implementations should be cheap and synchronous — the callback fires
-   * inside the watchdog `setTimeout` handler, before the abort cascade,
-   * and must never throw.
-   */
-  getActiveTurnSending?: () => boolean;
-}
 
 /**
  * Open an SSE connection to the assistant's events endpoint and emit typed
@@ -131,147 +77,18 @@ export function subscribeChatEvents(
   // The top-level cancel() targets whichever attempt is currently
   // active.
   let activeAbortController: AbortController | null = null;
-  let watchdogTimer: ReturnType<typeof setTimeout> | null = null;
   const requestedConversationId = conversationId ?? undefined;
-  // Cause of the most recent connect attempt teardown, consumed by the
-  // next reconnect when it invokes onReconnect. `"watchdog"` is set
-  // by armWatchdog before it aborts; left null otherwise so the
-  // reconnect path defaults to `"error"`.
-  let lastAbortCause: ChatStreamReconnectCause | null = null;
-  // Per-attempt liveness counters used to enrich the `sse_watchdog_fired`
-  // Sentry event with context about what (if anything) was flowing on
-  // the connection in the moments before the stall. These distinguish
-  // "no traffic at all since connect" (server never began responding)
-  // from "keepalives but no data" (vembda alive, daemon silent) from
-  // "data flowing but then stopped mid-turn" (mid-turn upstream death).
-  // Reset on every connect attempt so each watchdog fire reports the
-  // counts for its own attempt, not the cumulative session.
-  let lastSseAtMs: number | null = null;
-  let keepalivesReceivedSinceConnect = 0;
-  let dataFramesReceivedSinceConnect = 0;
 
-  const clearWatchdog = () => {
-    if (watchdogTimer) {
-      clearTimeout(watchdogTimer);
-      watchdogTimer = null;
-    }
-  };
-
-  // Reset (or arm) the idle watchdog for the supplied attempt. Called
-  // immediately before the for-await loop is entered and on every
-  // parsed SSE chunk thereafter — including heartbeat comments, which
-  // the SDK surfaces via onSseEvent even though they do not yield
-  // through the iterator. If no traffic arrives within idleTimeoutMs,
-  // abort the active fetch so the outer reconnect path runs; some
-  // runtimes (notably WKWebView on Capacitor iOS) hold a streaming
-  // fetch open at the network layer with no bytes flowing and no
-  // error surfaced to JavaScript, so server-side heartbeats are only
-  // a useful liveness signal if the client checks them.
-  const armWatchdog = (controller: AbortController) => {
-    if (cancelled) return;
-    clearWatchdog();
-    watchdogTimer = setTimeout(() => {
-      watchdogTimer = null;
-      lastAbortCause = "watchdog";
-      // Snapshot liveness state at the moment the watchdog fires.
-      // `wasTurnSending` is the single most useful aggregation
-      // dimension Sentry currently lacks: the 100% bucket=0
-      // reading collapses "benign idle stall after turn complete"
-      // and "user-harming stall during in-flight turn" into the
-      // same population, which makes the L2/L3 decision
-      // uninterpretable. Defensively wrap in try/catch because the
-      // caller-supplied snapshot is opaque to us; an exception here
-      // would prevent the diagnostic from being recorded.
-      let wasTurnSending: boolean | null = null;
-      try {
-        wasTurnSending = options.getActiveTurnSending?.() ?? null;
-      } catch {
-        // Diagnostics are best-effort and must never block recovery.
-      }
-      // `lastByteAgeMs` distinguishes "server never started responding
-      // to this attempt" (lastSseAtMs === null) from "some traffic
-      // arrived then stopped" (a positive age). With idleTimeoutMs
-      // = 45s a healthy heartbeat-driven idle stream sees ageMs
-      // ≤ the daemon's heartbeat interval just before the timer is
-      // reset, so an age > idleTimeoutMs at fire time would indicate
-      // the timer is running ahead of the bytes it's supposed to be
-      // resetting on — i.e. the SDK's onSseEvent callback is not
-      // firing for some frame shape. That alone would be a fix.
-      const lastByteAgeMs =
-        lastSseAtMs === null ? null : Date.now() - lastSseAtMs;
-      // Record before aborting so the diagnostic captures the
-      // attempt that actually stalled, even if the abort cascade
-      // tears state down before the next reconnect runs.
-      recordDiagnostic("sse_watchdog_fired", {
-        assistantId,
-        conversationId: requestedConversationId ?? null,
-        attempt: reconnectCount,
-        idleTimeoutMs,
-        wasTurnSending,
-        lastByteAgeMs,
-        keepalivesReceivedSinceConnect,
-        dataFramesReceivedSinceConnect,
-      });
-      // Mirror the same event into Sentry. sessionStorage events
-      // only ship off-device when a user manually attaches a
-      // diagnostics bundle, which biases the sample toward
-      // broken-and-noisy cases — exactly the cases the silent
-      // stall this watchdog detects is NOT. Sentry breadcrumbs
-      // attach to every subsequent error and captureMessage gives
-      // an aggregable count, so fleet-wide data answers the
-      // Layer 2 / Layer 3 question even when users never submit a
-      // bundle.
-      // https://docs.sentry.io/platforms/javascript/enriching-events/breadcrumbs/
-      Sentry.addBreadcrumb({
-        category: "sse.watchdog",
-        level: "warning",
-        message: "watchdog_fired",
-        data: {
-          assistantId,
-          attempt: reconnectCount,
-          idleTimeoutMs,
-          wasTurnSending,
-          lastByteAgeMs,
-          keepalivesReceivedSinceConnect,
-          dataFramesReceivedSinceConnect,
-        },
-      });
-      Sentry.captureMessage("sse_watchdog_fired", {
-        level: "warning",
-        // platform is the only fleet-wide signal that distinguishes
-        // Capacitor iOS from Safari iOS — Sentry's auto-detected
-        // os.name does not, but the idle-watchdog is iOS-only so the L2/L3
-        // decision needs the breakdown. tags are aggregable in
-        // Discover; extras are not. wasTurnSending is promoted to a
-        // tag so the user-harming vs benign split can be queried in
-        // one Discover groupBy without per-event drill-in.
-        // https://docs.sentry.io/product/explore/discover-queries/
-        // https://docs.sentry.io/concepts/key-terms/key-terms/#tags
-        tags: {
-          context: "sse_watchdog",
-          platform: resolvePlatformTag(),
-          attempt: String(reconnectCount),
-          wasTurnSending:
-            wasTurnSending === null ? "unknown" : String(wasTurnSending),
-        },
-        extra: {
-          assistantId,
-          conversationId: requestedConversationId ?? null,
-          attempt: reconnectCount,
-          idleTimeoutMs,
-          wasTurnSending,
-          lastByteAgeMs,
-          keepalivesReceivedSinceConnect,
-          dataFramesReceivedSinceConnect,
-        },
-      });
-      controller.abort();
-    }, idleTimeoutMs);
-  };
+  const watchdog = createStreamWatchdog({
+    idleTimeoutMs,
+    assistantId,
+    conversationId: requestedConversationId,
+    getActiveTurnSending: options.getActiveTurnSending,
+  });
 
   const cancel = () => {
     cancelled = true;
-    clearWatchdog();
+    watchdog.clear();
     activeAbortController?.abort();
   };
 
@@ -302,12 +119,8 @@ export function subscribeChatEvents(
     );
     // Reset per-attempt liveness counters so each watchdog fire
     // reports state for ITS attempt, not for the entire subscribe
-    // lifetime. lastSseAtMs stays null until the first SSE chunk
-    // arrives so the diagnostic can distinguish "server never
-    // responded" from "some traffic then stalled".
-    lastSseAtMs = null;
-    keepalivesReceivedSinceConnect = 0;
-    dataFramesReceivedSinceConnect = 0;
+    // lifetime.
+    watchdog.resetCounters();
     let streamError: Error | null = null;
     try {
       // The wire-field gate prefers `conversationId` on daemons that
@@ -341,35 +154,21 @@ export function subscribeChatEvents(
         onSseEvent: (event) => {
           // Fires for every parsed SSE chunk including heartbeat
           // comments (which the SDK surfaces with `data === undefined`
-          // because comment frames have no `data:` line). Hooking
-          // the watchdog reset here is what makes the timeout safe
-          // in the foreground: a healthy idle connection still
-          // receives a heartbeat every ~30s between user turns and
-          // will not be force-reconnected.
-          //
-          // Counting heartbeats vs data frames separately lets the
-          // watchdog diagnostic distinguish "vembda alive, daemon
-          // silent" (keepalives > 0, dataFrames = 0) from
-          // "stream died mid-turn" (keepalives ≥ 0, dataFrames > 0)
-          // from "server never started responding" (both 0).
-          // event.data is typed as TData but is undefined at runtime
-          // for comment-only chunks per @hey-api SDK semantics; the
-          // cast lets the runtime check stay precise without
-          // relying on a type assertion at every read.
-          if (typeof (event as { data?: unknown }).data === "undefined") {
-            keepalivesReceivedSinceConnect++;
-          } else {
-            dataFramesReceivedSinceConnect++;
+          // because comment frames have no `data:` line).
+          const isData = typeof (event as { data?: unknown }).data !== "undefined";
+          if (isData) {
             markClientEstablished(sseDebugClientId);
           }
-          lastSseAtMs = Date.now();
-          armWatchdog(abortController);
+          watchdog.recordTraffic(isData);
+          if (!cancelled) {
+            watchdog.arm(abortController, reconnectCount);
+          }
         },
       });
 
       if (isReconnect && !cancelled) {
-        const cause: ChatStreamReconnectCause = lastAbortCause ?? "error";
-        lastAbortCause = null;
+        const cause: ChatStreamReconnectCause =
+          watchdog.consumeLastAbortCause() ?? "error";
         try {
           await options.onReconnect?.(cause);
         } catch {
@@ -385,7 +184,9 @@ export function subscribeChatEvents(
       // take several seconds. Arming the timer earlier would charge
       // that reconcile time against idleTimeoutMs and could abort
       // the new attempt before any SSE traffic ever started.
-      armWatchdog(abortController);
+      if (!cancelled) {
+        watchdog.arm(abortController, reconnectCount);
+      }
 
       let receivedEvent = false;
 
@@ -398,7 +199,7 @@ export function subscribeChatEvents(
           // has already covered this chunk, but resetting again here
           // wires the watchdog independently of the SDK's internal
           // callback ordering.
-          armWatchdog(abortController);
+          watchdog.arm(abortController, reconnectCount);
 
           const data = typeof payload === "string"
             ? (() => {
@@ -447,7 +248,7 @@ export function subscribeChatEvents(
         // teardown that happens close to the idle deadline lets the
         // timer run during the reconnect backoff and falsely set
         // lastAbortCause = "watchdog" on a recoverable error path.
-        clearWatchdog();
+        watchdog.clear();
       }
       if (cancelled) {
         return;
