+++ apps/web/src/lib/streaming/stream-watchdog.ts
@@ -1,0 +1,208 @@
+/**
+ * Idle watchdog for SSE stream connections.
+ *
+ * Detects silently stalled connections — notably on iOS WKWebView
+ * (Capacitor), where the fetch promise can hold a streaming connection
+ * open at the network layer with no bytes flowing and no error surfaced
+ * to JavaScript. The daemon emits a heartbeat comment every ~30 s; this
+ * watchdog aborts the active fetch when no SSE traffic (events OR
+ * heartbeat comments) arrives within a configurable window.
+ *
+ * Telemetry is recorded to both sessionStorage diagnostics (via
+ * {@link recordDiagnostic}) and Sentry (breadcrumb + captureMessage)
+ * so fleet-wide stall frequency can be measured without relying on
+ * user-submitted diagnostic bundles.
+ *
+ * @see https://docs.sentry.io/platforms/javascript/enriching-events/breadcrumbs/
+ * @see https://docs.sentry.io/product/explore/discover-queries/
+ */
+
+import * as Sentry from "@sentry/browser";
+
+import { recordDiagnostic, resolvePlatformTag } from "@/lib/diagnostics";
+import type { ChatStreamReconnectCause } from "@/lib/streaming/stream-transport-types";
+
+// ---------------------------------------------------------------------------
+// Config
+// ---------------------------------------------------------------------------
+
+export interface StreamWatchdogConfig {
+  /** Milliseconds of silence before the watchdog fires. */
+  idleTimeoutMs: number;
+  assistantId: string;
+  conversationId: string | undefined;
+  /**
+   * Snapshot whether the caller-owned turn state machine is currently
+   * sending. Forwarded to Sentry as the `wasTurnSending` tag so
+   * user-harming stalls (during an in-flight turn) can be distinguished
+   * from benign ones (idle stream after a turn completed).
+   */
+  getActiveTurnSending?: () => boolean;
+}
+
+// ---------------------------------------------------------------------------
+// Public API
+// ---------------------------------------------------------------------------
+
+export interface StreamWatchdog {
+  /**
+   * Reset (or arm) the idle timer for the given controller. Called
+   * immediately before the for-await loop and on every parsed SSE
+   * chunk — including heartbeat comments. If no traffic arrives
+   * within `idleTimeoutMs`, the controller is aborted.
+   */
+  arm(controller: AbortController, attempt: number): void;
+  /** Cancel any pending timer. */
+  clear(): void;
+  /** Reset per-attempt liveness counters (call on each new connect). */
+  resetCounters(): void;
+  /**
+   * Record an SSE traffic event. `isData` distinguishes data frames
+   * (yielded through the iterator) from heartbeat comment frames
+   * (surfaced via `onSseEvent` with `data === undefined`).
+   */
+  recordTraffic(isData: boolean): void;
+  /**
+   * Return and clear the cause set by the most recent watchdog fire.
+   * Returns `null` if the watchdog has not fired since the last
+   * consume (or ever).
+   */
+  consumeLastAbortCause(): ChatStreamReconnectCause | null;
+}
+
+/**
+ * Create a watchdog instance that monitors SSE stream liveness.
+ *
+ * The returned handle encapsulates the timer, liveness counters, and
+ * Sentry telemetry. The transport calls {@link StreamWatchdog.arm}
+ * on every SSE chunk and {@link StreamWatchdog.clear} when the
+ * read loop exits.
+ */
+export function createStreamWatchdog(
+  config: StreamWatchdogConfig,
+): StreamWatchdog {
+  const { idleTimeoutMs, assistantId, conversationId, getActiveTurnSending } =
+    config;
+
+  let timer: ReturnType<typeof setTimeout> | null = null;
+  let lastSseAtMs: number | null = null;
+  let keepalivesReceivedSinceConnect = 0;
+  let dataFramesReceivedSinceConnect = 0;
+  let lastAbortCause: ChatStreamReconnectCause | null = null;
+
+  const clear = () => {
+    if (timer) {
+      clearTimeout(timer);
+      timer = null;
+    }
+  };
+
+  const arm = (controller: AbortController, attempt: number) => {
+    clear();
+    timer = setTimeout(() => {
+      timer = null;
+      lastAbortCause = "watchdog";
+
+      // Snapshot turn-sending state at the moment the watchdog fires.
+      // This is the single most useful Sentry aggregation dimension:
+      // it separates user-harming stalls (in-flight turn) from benign
+      // ones (idle stream after turn complete). Defensively wrapped
+      // because the caller-supplied snapshot is opaque.
+      let wasTurnSending: boolean | null = null;
+      try {
+        wasTurnSending = getActiveTurnSending?.() ?? null;
+      } catch {
+        // Diagnostics are best-effort and must never block recovery.
+      }
+
+      // Distinguish "server never started responding" (null) from
+      // "some traffic arrived then stopped" (positive age).
+      const lastByteAgeMs =
+        lastSseAtMs === null ? null : Date.now() - lastSseAtMs;
+
+      recordDiagnostic("sse_watchdog_fired", {
+        assistantId,
+        conversationId: conversationId ?? null,
+        attempt,
+        idleTimeoutMs,
+        wasTurnSending,
+        lastByteAgeMs,
+        keepalivesReceivedSinceConnect,
+        dataFramesReceivedSinceConnect,
+      });
+
+      // Mirror into Sentry for fleet-wide aggregation.
+      // sessionStorage diagnostics only ship when a user manually
+      // attaches a bundle, biasing toward broken-and-noisy cases.
+      // Sentry breadcrumbs + captureMessage give an aggregable count
+      // across all users.
+      // @see https://docs.sentry.io/platforms/javascript/enriching-events/breadcrumbs/
+      Sentry.addBreadcrumb({
+        category: "sse.watchdog",
+        level: "warning",
+        message: "watchdog_fired",
+        data: {
+          assistantId,
+          attempt,
+          idleTimeoutMs,
+          wasTurnSending,
+          lastByteAgeMs,
+          keepalivesReceivedSinceConnect,
+          dataFramesReceivedSinceConnect,
+        },
+      });
+      Sentry.captureMessage("sse_watchdog_fired", {
+        level: "warning",
+        // `platform` distinguishes Capacitor iOS from Safari iOS —
+        // Sentry's auto-detected os.name does not, but the watchdog
+        // is iOS-specific so the L2/L3 decision needs the breakdown.
+        // `wasTurnSending` is promoted to a tag so the user-harming
+        // vs benign split is queryable in Discover without per-event
+        // drill-in.
+        // @see https://docs.sentry.io/concepts/key-terms/key-terms/#tags
+        tags: {
+          context: "sse_watchdog",
+          platform: resolvePlatformTag(),
+          attempt: String(attempt),
+          wasTurnSending:
+            wasTurnSending === null ? "unknown" : String(wasTurnSending),
+        },
+        extra: {
+          assistantId,
+          conversationId: conversationId ?? null,
+          attempt,
+          idleTimeoutMs,
+          wasTurnSending,
+          lastByteAgeMs,
+          keepalivesReceivedSinceConnect,
+          dataFramesReceivedSinceConnect,
+        },
+      });
+
+      controller.abort();
+    }, idleTimeoutMs);
+  };
+
+  const resetCounters = () => {
+    lastSseAtMs = null;
+    keepalivesReceivedSinceConnect = 0;
+    dataFramesReceivedSinceConnect = 0;
+  };
+
+  const recordTraffic = (isData: boolean) => {
+    if (isData) {
+      dataFramesReceivedSinceConnect++;
+    } else {
+      keepalivesReceivedSinceConnect++;
+    }
+    lastSseAtMs = Date.now();
+  };
+
+  const consumeLastAbortCause = (): ChatStreamReconnectCause | null => {
+    const cause = lastAbortCause;
+    lastAbortCause = null;
+    return cause;
+  };
+
+  return { arm, clear, resetCounters, recordTraffic, consumeLastAbortCause };
+}
